* 'dev' of http://172.20.0.22/datamiddle/backend/qk-dam:
  [数据标准] 码表英文表名称设置唯一属性;标准增加排序功能;

+++ qk-dam/qike-dm-datastandards/src/main/java/com/qk/dm/datastandards/repositories/DsdBasicinfoRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface DsdBasicinfoRepository extends BaseRepository<DsdBasicinfo, Integer> {
+public interface DsdBasicinfoRepository extends BaseRepository<DsdBasicinfo, Long> {
 
-  @Modifying
-  @Query(" UPDATE DsdBasicinfo set dsdLevel = :dsdDirLevel where dsdLevelId = :dirDsdId ")
-  void updateDirLevelByDirId(
-      @Param("dsdDirLevel") String dsdDirLevel, @Param("dirDsdId") String dirDsdId);
+    @Modifying
+    @Query(" UPDATE DsdBasicinfo set dsdLevel = :dsdDirLevel where dsdLevelId = :dirDsdId ")
+    void updateDirLevelByDirId(@Param("dsdDirLevel") String dsdDirLevel, @Param("dirDsdId") String dirDsdId);
+
 }
